+++ src/algorithms/imageProcessing/features/HCPT.java
@@ -233,7 +233,7 @@ public class HCPT {
             //System.out.println(" " + yA + " -- " + yB + " sum="+sum + ", " + sumA + "," + sumB);
         }
         
-        float sim = sum / ((float)Math.min(sumA, sumB));
+        float sim = sum / Math.min(sumA, sumB);
         
         return sim;
     }
